+++ workshop_final_project/src/rag_app/chat_model.py
@@ -1,0 +1,46 @@
+import time
+import streamlit as st
+
+from .ai_client import AIProjectClientDefinition
+
+
+def safe_gpt_call(func, *args, max_retries=5, wait_seconds=60, **kwargs):
+    for attempt in range(max_retries):
+        try:
+            return func(*args, **kwargs)
+        except Exception as e:
+            if hasattr(e, "status_code") and e.status_code == 429:
+                st.warning(
+                    f"Rate limit Azure superato! Aspetto {wait_seconds} secondi... (Tentativo {attempt+1}/{max_retries})"
+                )
+                time.sleep(wait_seconds)
+            else:
+                raise e
+    raise Exception("Superato il numero massimo di retry per il rate limit.")
+
+
+class ChatCompletionModel(AIProjectClientDefinition):
+    def __init__(self, model_name: str = "gpt-4o") -> None:
+        super().__init__()
+        self.model_name = model_name
+        self.azure_client = self.client.inference.get_azure_openai_client(
+            api_version="2025-01-01-preview"
+        )
+
+    def ask_about_document(self, content: str, question: str) -> str:
+        messages = [
+            {
+                "role": "system",
+                "content": "Sei un assistente AI specializzato in analisi di documenti testuali.",
+            },
+            {"role": "user", "content": f"Documento:\n{content}\n\nDomanda: {question}"},
+        ]
+        response = safe_gpt_call(
+            self.azure_client.chat.completions.create,
+            model=self.model_name,
+            messages=messages,
+            max_tokens=512,
+            temperature=0.7,
+            top_p=1.0,
+        )
+        return response.choices[0].message.content
